Breakup ExecutionEventJsonPrinter
