+++ test/shared/src/main/scala/zio/test/results/ExecutionEventJsonPrinter.scala
@@ -1,102 +1,58 @@
 package zio.test.results
 
-import zio.test.{
-  ExecutionEvent,
-  ExecutionEventPrinter,
-  TestAnnotationMap,
-  TestAnnotationRenderer,
-  TestFailure,
-  TestSuccess
-}
-import zio.{Scope, Trace, ZIO, ZLayer, ZOutputStream}
+import zio.test._
+import zio._
 
-import java.io
 import java.io.IOException
-import java.nio.file.Path
 
 object ExecutionEventJsonPrinter {
   val live = {
     implicit val trace = Trace.empty
-    val instance       = new Live()
+    val instance       = new Live("target/test-reports-zio/output.json", Json)
 
     ZLayer.scoped(
       ZIO.acquireRelease(
-        instance.reset("target/test-reports-zio/output.json") *>
+        instance.resetContents("[") *>
           instance.makeOutputDirectory().orDie *>
           ZIO.succeed(instance)
       )(_ =>
-        instance.removeTrailingComma("target/test-reports-zio/output.json") *>
-          instance.closeJsonStructure("target/test-reports-zio/output.json") *>
-          ZIO.debug("Should remove last command and add closing square bracket")
+        instance.removeTrailingComma() *>
+          instance.closeJsonStructure()
       )
     )
   }
 
-  class Live() extends ExecutionEventPrinter {
+  class Live(resultPath: String, serializer: ResultSerializer) extends ExecutionEventPrinter {
     override def print(event: ExecutionEvent): ZIO[Any, Nothing, Unit] = {
       implicit val trace = Trace.empty
-      writeFile("target/test-reports-zio/output.json", jsonify(event)).orDie
+      writeFile(resultPath, serializer.render(event)).orDie
     }
 
-    private def jsonify[E](test: Either[TestFailure[E], TestSuccess]): String =
-      test match {
-        case Left(value) =>
-          "Failure"
-        case Right(value) =>
-          "Success"
-      }
-
-    private def jsonify(testAnnotationMap: TestAnnotationMap): String =
-      TestAnnotationRenderer.default
-        .run(List.empty, testAnnotationMap)
-        .map(s => s.replace("\"", "\\\""))
-        .mkString(" : ")
-
-    private def jsonify(executionEvent: ExecutionEvent): String = executionEvent match {
-      case ExecutionEvent.Test(labelsReversed, test, annotations, ancestors, duration, id, fullyQualifiedName) =>
-        s"""
-           |  {
-           |     "name" : "$fullyQualifiedName/${labelsReversed.reverse.mkString("/")}",
-           |     "status" : "${jsonify(test)}",
-           |     "durationMillis" : "${duration}",
-           |     "annotations" : "${jsonify(annotations)}"
-           |  },""".stripMargin
-      case ExecutionEvent.SectionStart(labelsReversed, id, ancestors) =>
-        ""
-      case ExecutionEvent.SectionEnd(labelsReversed, id, ancestors) =>
-        // TODO Deal with trailing commas
-        ""
-      case ExecutionEvent.TopLevelFlush(id) => "TODO TopLevelFlush"
-      case ExecutionEvent.RuntimeFailure(id, labelsReversed, failure, ancestors) =>
-        "TODO RuntimeFailure"
-    }
-
-    def reset(path: => String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
+    def resetContents(newContents: String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
       ZIO
-        .acquireReleaseWith(ZIO.attemptBlockingIO(new java.io.FileWriter(path, false)))(f =>
+        .acquireReleaseWith(ZIO.attemptBlockingIO(new java.io.FileWriter(resultPath, false)))(f =>
           ZIO.attemptBlocking(f.close()).orDie
         ) { f =>
-          ZIO.attemptBlockingIO(f.write("["))
+          ZIO.attemptBlockingIO(f.write(newContents))
         }
         .orDie
 
-    def closeJsonStructure(path: => String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
-      writeFile(path, "]").orDie
+    def closeJsonStructure()(implicit trace: Trace): ZIO[Any, Nothing, Unit] =
+      writeFile(resultPath, "]").orDie
 
     def makeOutputDirectory()(implicit trace: Trace) = ZIO.attempt {
       import java.nio.file.{Files, Paths}
 
-      val fp = Paths.get("target/test-reports-zio/newfile.txt")
+      val fp = Paths.get(resultPath)
       Files.createDirectories(fp.getParent)
-//      Files.createFile(fp)
 
     }
 
     import java.io._
 
-    def removeTrailingComma(filePath: String)(implicit trace: Trace): ZIO[Any, Nothing, Unit] = ZIO.succeed {
+    def removeTrailingComma()(implicit trace: Trace): ZIO[Any, Nothing, Unit] = ZIO.succeed {
       try {
-        val file = new RandomAccessFile(filePath, "rw")
+        val file = new RandomAccessFile(resultPath, "rw")
         // Move the file pointer to the last character
         file.seek(file.length - 1)
         // Read backwards from the end of the file until we find a non-whitespace character
@@ -125,27 +81,9 @@ object ExecutionEventJsonPrinter {
       ZIO.attemptBlockingIO(f.append(content))
     }
 
-  def writeFile(path: => Path, content: => String)(implicit
-    trace: Trace,
-    d: DummyImplicit
-  ): ZIO[Any, IOException, Unit] =
-    writeFile(path.toString, content)
-
-  def writeFileOutputStream(
-    path: => String
-  )(implicit trace: Trace): ZIO[Scope, IOException, ZOutputStream] =
-    ZIO
-      .acquireRelease(
-        ZIO.attemptBlockingIO {
-          val fos = new io.FileOutputStream(path)
-          (fos, ZOutputStream.fromOutputStream(fos))
-        }
-      )(tuple => ZIO.attemptBlocking(tuple._1.close()).orDie)
-      .map(_._2)
-
-  def writeFileOutputStream(
-    path: => Path
-  )(implicit trace: Trace, d: DummyImplicit): ZIO[Scope, IOException, ZOutputStream] =
-    writeFileOutputStream(path.toString)
 
 }
+
+
+
+
