+++ test/shared/src/main/scala/zio/test/results/ResultSerializer.scala
@@ -1,0 +1,7 @@
+package zio.test.results
+import zio.test.ExecutionEvent
+import zio.test.results.Json.jsonify
+
+trait ResultSerializer {
+  def render(executionEvent: ExecutionEvent): String
+}
